requirements.txt
apispec==6.3.0
blinker==1.6.2
cfgv==3.4.0
click==8.1.7
distlib==0.3.7
filelock==3.12.4
Flask==2.3.3
flask-blueprint==1.3.0
Flask-JWT-Extended==4.5.2
flask-rest-api==0.17.1
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
identify==2.5.29
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
nodeenv==1.8.0
packaging==23.1
platformdirs==3.10.0
pre-commit==3.4.0
PyJWT==2.8.0
python-decouple==3.8
PyYAML==6.0.1
SQLAlchemy==2.0.21
typing_extensions==4.8.0
virtualenv==20.24.5
webargs==8.3.0
Werkzeug==2.3.7
